✏️ update : 전체 고객, 특정 그룹 고객 반경 검색 api 연동 코드 수정 완료
전체 고객, 특정 그룹 고객 반경 검색 api 연동 코드 수정 완료, 검색창 구현 진행중

> Related to : #53

+++ app/src/main/java/com/example/gajamap/data/repository/RadiusRepository.kt
@@ -7,8 +7,8 @@ class RadiusRepository {
     private val radius = GajaMapApplication.sRetrofit.create(RadiusInterface::class.java)
 
     // 전체 고객 대상 반경 검색
-    suspend fun wholeRadius(radius1: Double, latitude: Double, longitude: Double) = radius.wholeRadius(radius1, latitude, longitude)
+    suspend fun wholeRadius(radius1: Int, latitude: Double, longitude: Double) = radius.wholeRadius(radius1, latitude, longitude)
 
     // 특정 그룹 내에 고객 반경 검색
-    suspend fun specificRadius(groupId : Long, radius1: Double, latitude: Double, longitude: Double) = radius.specificRadius(groupId, radius1, latitude, longitude)
+    suspend fun specificRadius(groupId : Long, radius1: Int, latitude: Double, longitude: Double) = radius.specificRadius(groupId, radius1, latitude, longitude)
 }
