+++ app/src/main/java/com/example/gajamap/data/service/RadiusInterface.kt
@@ -1,5 +1,6 @@
 package com.example.gajamap.data.service
 
+import com.example.gajamap.data.model.GetRadiusResponse
 import com.example.gajamap.data.model.RadiusResponse
 import retrofit2.Response
 import retrofit2.http.GET
@@ -9,9 +10,9 @@ import retrofit2.http.Query
 interface RadiusInterface {
     // 전체 고객 대상 반경 검색
     @GET("/api/clients/nearby")
-    suspend fun wholeRadius(@Query("radius") radius: Double, @Query("latitude") latitude: Double, @Query("longitude") longitude: Double) : Response<RadiusResponse>
+    suspend fun wholeRadius(@Query("radius") radius: Int, @Query("latitude") latitude: Double, @Query("longitude") longitude: Double) : Response<RadiusResponse>
 
-    // 전체 고객 대상 반경 검색
+    // 그룹 고객 대상 반경 검색
     @GET("/api/group/{groupId}/clients/nearby")
-    suspend fun specificRadius(@Path("groupId") groupId : Long, @Query("radius") radius: Double, @Query("latitude") latitude: Double, @Query("longitude") longitude: Double) : Response<RadiusResponse>
+    suspend fun specificRadius(@Path("groupId") groupId : Long, @Query("radius") radius: Int, @Query("latitude") latitude: Double, @Query("longitude") longitude: Double) : Response<RadiusResponse>
 }
